Refactor UserResponseDTO
