+++ marcus-bike-api/src/main/java/com/marcusbike/marcus_bike_api/services/UserService.java
@@ -5,7 +5,7 @@ import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 
-import com.marcusbike.marcus_bike_api.dto.UserDTO;
+import com.marcusbike.marcus_bike_api.dto.response.UserResponseDTO;
 import com.marcusbike.marcus_bike_api.mappers.UserMapper;
 import com.marcusbike.marcus_bike_api.repositories.UserRepository;
 
@@ -17,7 +17,7 @@ public class UserService {
 
     private final UserRepository userRepository;
 
-    public List<UserDTO> findAll() {
+    public List<UserResponseDTO> findAll() {
         return userRepository.findAll().stream().map(user -> UserMapper.toDTO(user)).collect(Collectors.toList());
     }
